+++ crates/bz-lib/src/compression/mod.rs
@@ -1,9 +1,10 @@
 pub mod archive;
 mod compress;
 mod decompress;
+mod extract;
 pub mod streams;
 
-use crate::cli::{CompressConfig, DecompressConfig};
+use crate::cli::{CompressConfig, DecompressConfig, ExtractConfig};
 use anyhow::Result;
 
 pub fn compress(config: &CompressConfig) -> Result<()> {
@@ -12,4 +13,8 @@ pub fn compress(config: &CompressConfig) -> Result<()> {
 
 pub fn decompress(config: &DecompressConfig) -> Result<()> {
     decompress::decompress(config)
+}
+
+pub fn extract(config: &ExtractConfig) -> Result<()> {
+    extract::extract(config)
 }