+++ crates/bz-lib/src/lib.rs
@@ -2,5 +2,5 @@ pub mod cli;
 pub mod compression;
 pub mod io;
 
-pub use cli::{CompressConfig, DecompressConfig};
-pub use compression::{compress, decompress};
+pub use cli::{CompressConfig, DecompressConfig, ExtractConfig};
+pub use compression::{compress, decompress, extract};